% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deflation}
\alias{deflation}
\title{Deflation}
\usage{
deflation(X, d)
}
\arguments{
\item{X}{a numeric matrix}

\item{d}{a vector to deflate}
}
\value{
Y a deflated matrix
}
\description{
To deflate a matrix
}
\keyword{internal}

