\name{iteraite_fun_Iter}
\alias{iteraite_fun_Iter}
\title{Iteration function}
\usage{
  iteraite_fun_Iter(tm, K.Data, concat.Data, K, M, P, n, N)
}
\arguments{
  \item{The}{parameters in the main functions}
}
\value{
  The results for each iteration
}
\description{
  To calculate outputs in each iteration
}
\keyword{internal}

