% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadingsplot.R
\name{loadingsplot}
\alias{loadingsplot}
\title{loadings plot}
\usage{
loadingsplot(x, axes = c(1, 2), INERTIE = NULL, cex = NULL,
  font.lab = NULL)
}
\arguments{
\item{x}{results of the proposed multigroup methods in the package}

\item{axes}{a vector of two selected components}

\item{INERTIE}{if there is information about inertia}

\item{cex}{character expansion for text by default .85}

\item{font.lab}{type of font by default 3}
}
\value{
loadings plot
}
\description{
plots of variables (loadings)
}
\examples{
Data = iris[,-5]
Group = iris[,5]
res.mgPCA = mgPCA (Data, Group, graph=TRUE)
loadingsplot(res.mgPCA, axes=c(1,2))
}

