% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multigroup.R
\docType{package}
\name{multigroup}
\alias{multigroup}
\alias{multigroup-package}
\title{multigroup: methods for multigroup data analysis}
\description{
This package includes several methods to study multigroup data,
where the same set of variables are measured on different groups of
individuals.
}
\section{Some Functions}{


multigroup provides a set of functions for multigroup analysis:

\itemize{
  \item \code{\link{BGC}}: Between Group Comparison
  \item \code{\link{DCCSWA}}: Dual Common Component and Specific Weights Analysis
  \item \code{\link{DGPA}}: Dual Generalized Procrustes Analysis
  \item \code{\link{DSTATIS}}: Dual STATIS
  \item \code{\link{FCPCA}}: Flury's Common Principal Component Analysis
  \item \code{\link{mgPCA}}:Multigroup Principal Component Analysis
  \item \code{\link{mgPLS}}: Multigroup Partial Least Squares Regression
  \item \code{\link{mbmgPCA}}: Multiblock and multigroup PCA
}
}

