% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normM.R
\name{normM}
\alias{normM}
\title{Normalize function for a matrix}
\usage{
normM(X)
}
\arguments{
\item{X}{a numeric matrix}
}
\value{
a matrix with norm 1
}
\description{
To calculate normmalize of a matrix
}
\keyword{internal}
