% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normv.R
\name{normv}
\alias{normv}
\title{Normalize function for a vector}
\usage{
normv(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector with norm 1
}
\description{
To calculate normalize of a vector
}
\keyword{internal}

