\name{oliveoil}

\alias{oliveoil}

\docType{data}

\title{Sensory and physico-chemical data of olive oils}

\description{
  A data set with scores on 6 attributes from a sensory panel and measurements of 5 physico-chemical quality parameters on 16 olive oil samples. The first five oils are Greek, the next five are Italian and the last six are Spanish (Package pls).
}

\usage{data(oliveoil)}

\format{
  A data frame with 16 observations on the following 2 variables.
  sensory a matrix with 6 columns. Scores for attributes yellow, green, brown, glossy, transp, and syrup. chemical a matrix with 5 columns. Measurements of acidity, peroxide, K232, K270, and DK (Package pls).
}

\source{
  Package pls
}

\keyword{datasets}