% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.mg.R
\name{plot.mg}
\alias{plot.mg}
\title{Plots for multigroup objects}
\usage{
\method{plot}{mg}(x, axes = c(1, 2), cex = NULL, font.lab = NULL, ...)
}
\arguments{
\item{x}{results of multigroup method in the package}

\item{axes}{by default the first two components}

\item{cex}{character expansion for text by default .85}

\item{font.lab}{type of font by default 3}

\item{\dots}{Further arguments are ignored}
}
\value{
loadings and scores plots
}
\description{
plots of variables (loadings) and individuals (scores) if TRUE
}

