% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project operatior}
\usage{
project(X)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
To calculate projection of a matrix
}
\keyword{internal}

