% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_function}
\alias{similarity_function}
\title{similarity function}
\usage{
similarity_function(loadings_matrices, NAMES)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
To calculate similarity among common and group loadings
}
\keyword{internal}

