% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deflation.R
\name{deflation}
\alias{deflation}
\title{Deflation}
\usage{
deflation(X, d)
}
\arguments{
\item{X}{a numeric matrix}

\item{d}{a vector to deflate}
}
\value{
Y a deflated matrix
}
\description{
To deflate a matrix
}
\keyword{internal}

