% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_initial_Iter.R
\name{select_initial_Iter}
\alias{select_initial_Iter}
\title{Selection of initial values}
\usage{
select_initial_Iter(tm = tm, K.Data = K.Data, concat.Data = concat.Data,
  K = K, M = M, P = P, n = n, N = N)
}
\arguments{
\item{X}{a numeric matrix}

\item{parametrs}{from iteraite.fun.Iter function}
}
\value{
maximum of the optimal criterion
}
\description{
To Select the initial values
}
\keyword{internal}

