% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity_noncum}
\alias{similarity_noncum}
\title{Non cumulative similarity function}
\usage{
similarity_noncum(loadings_matrices, NAMES)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
To calculate non cumulative similarity among common and group loadings
}
\keyword{internal}

