% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCsimilarity.R
\name{similarity_noncum}
\alias{similarity_noncum}
\title{Non cumulative similarity function}
\usage{
similarity_noncum(loadings_matrices, NAMES)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
To calculate non cumulative similarity among common and group loadings
}
\keyword{internal}

