% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transfer_Group.R
\name{transfer_Group}
\alias{transfer_Group}
\title{Transter function}
\usage{
transfer_Group(Data, Group)
}
\arguments{
\item{Data}{a numeric (quantitative) matrix or data frame}

\item{Group}{a vector of factors associated with group structure}
}
\value{
list with the following results:

\item{Data.group}{original data}
}
\description{
Transfer data to M groups of individuals
}
\seealso{
\code{\link{split}}, \code{\link{iris}}
}
\keyword{internal}

